Move numeric_bounds to base module
